+++ Infrastructure/utilities/block-chain.go
@@ -1,4 +1,4 @@
-package services
+package utilities
 
 import (
 	ecdsa "crypto/ecdsa"
