+++ Infrastructure/utilities/key-generator.go
@@ -1,4 +1,4 @@
-package services
+package utilities
 
 import (
 	ecdsa "crypto/ecdsa"
