+++ server.go
@@ -4,9 +4,10 @@ import (
 	commands "bitshare-chain/application/commands"
 	mongo_context "bitshare-chain/application/data-access/context"
 	repositories "bitshare-chain/application/data-access/repositories"
+	services "bitshare-chain/application/services"
 	validation "bitshare-chain/application/validation"
 	settings "bitshare-chain/infrastructure/settings"
-	services "bitshare-chain/infrastructure/utilities"
+	utilities "bitshare-chain/infrastructure/utilities"
 	controllers "bitshare-chain/web/controllers"
 
 	gin "github.com/gin-gonic/gin"
@@ -19,7 +20,7 @@ func main() {
 	//DB
 	dbOptions := settings.MongoDbOptions{
 		DatabaseName:     "GoBitshareChain",
-		ConnectionString: "mongodb://root:rootpassword@go-bitshare-mongodb:27017/?authMechanism=SCRAM-SHA-256",
+		ConnectionString: "mongodb://root:rootpassword@go-bitshare-mongodb:27000/?authMechanism=SCRAM-SHA-256",
 	}
 
 	mongoContext, err := mongo_context.NewMongoContext(&dbOptions)
@@ -30,9 +31,11 @@ func main() {
 
 	//REPOS
 	walletAccountRepository := repositories.NewWalletAccountRepository(mongoContext)
+	nodeMetadataRepository := repositories.NewNodeMetadataRepository(mongoContext)
 
 	//SERVICES
-	keyGenerator := &services.KeyGenerator{}
+	keyGenerator := &utilities.KeyGenerator{}
+	metadataService := services.NewMetadataService(*keyGenerator, *nodeMetadataRepository)
 
 	//VALIDATOR
 	validator := validation.NewValidator()
@@ -42,7 +45,7 @@ func main() {
 	testHandler := commands.NewTestCommandHandler(validator)
 
 	//CONTROLLERS
-	chainController := controllers.NewChainController(ginRouter, createWalletAccountCommandHandler)
+	chainController := controllers.NewChainController(ginRouter, createWalletAccountCommandHandler, metadataService)
 	chainController.SetupChainController()
 
 	testController := controllers.NewTestController(ginRouter, testHandler)
